requirements.txt
discord
python-dotenv
numpy
